New caught route and re-organized to work with SSE streaming.
