+++ requirements.txt
@@ -78,6 +78,7 @@ soupsieve==1.9.1
 sqlacodegen==2.0.1
 SQLAlchemy==1.3.5
 sqlalchemy-stubs==0.1
+sseclient==0.0.24
 synphot==0.1.3
 traitlets==4.3.2
 typed-ast==1.3.1
