+++ src/controllers/__init__.py
@@ -7,7 +7,7 @@ Flask-RESTPlus namespacing example: https://flask-restplus.readthedocs.io/en/sta
 # import os
 import logging
 import traceback
-from typing import Any, Tuple, Optional
+from typing import Any, Tuple
 from flask_restplus import Api
 from flask import jsonify, Blueprint
 from flask.wrappers import Response
@@ -18,12 +18,16 @@ from env import ENV, EDeploymentEnvironment as EDE
 from .demo_routes import API as demos
 from .query import API as query
 from .images import API as images
+from .stream import API as stream
+from .caught import API as caught
 
 logger: logging.Logger = logging.getLogger(__name__)
 logger.info('"<><><> IMPORTING CONTROLLERS <><><>"')
 
 
 # Choose port to run app locally based on deployment environment
+URL_PREFIX: str
+TITLE_SUFFIX: str
 if ENV.DEPLOYMENT_ENV == EDE.PROD:
     URL_PREFIX = '/catch'
     TITLE_SUFFIX = ''
@@ -54,7 +58,9 @@ REST_PLUS_APIS = Api(
 
 # Combine Namespaces
 REST_PLUS_APIS.add_namespace(query)
+REST_PLUS_APIS.add_namespace(caught)
 REST_PLUS_APIS.add_namespace(images)
+REST_PLUS_APIS.add_namespace(stream)
 REST_PLUS_APIS.add_namespace(demos)
 
 # Add error handlers:
